requirements.txt
google-cloud-storage==2.7.0
smart-open==6.3.0